backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-administration==4.5.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dateparser==1.2.1
Django==5.1.6
django-mssql-backend==2.8.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
packaging==24.2
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyodbc==5.2.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
